+++ CHANGELOG.md
@@ -12,6 +12,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Option to specify custom CA certificate resources for TLS configuration.
 - Ensure all Certificate resources have `dnsNames` set.
 
+### Fixed
+
+- The operator will no longer restart reconciliation on every change to `ServiceAccount` resources.
+
 ## [v2.5.1] - 2024-05-02
 
 ### Added