+++ CHANGELOG.md
@@ -5,6 +5,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
 ## [v2.5.2] - 2024-07-17
 
 ### Added
@@ -875,3 +877,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [v2.5.0]: https://github.com/piraeusdatastore/piraeus-operator/compare/v2.4.1...v2.5.0
 [v2.5.1]: https://github.com/piraeusdatastore/piraeus-operator/compare/v2.5.0...v2.5.1
 [v2.5.2]: https://github.com/piraeusdatastore/piraeus-operator/compare/v2.5.1...v2.5.2
+[Unreleased]: https://github.com/piraeusdatastore/piraeus-operator/compare/v2.5.2...HEAD